Add files via upload
Implemented the following proxies to facilitate ordering:

SupplierProxy.java: This class implements the logic of the restock operation by forwarding the request to the restock operation of the Facade. If the request is a placeOrder operation request then it forwards the request to the next proxy in the sequence (i.e. the LowQuantityProxy) .  

LowQuantityProxy.java. This class implements the logic of the placeOrder operation by forwarding the request to the placeOrder operation of the Facade. If the request is a placeOrder operation request and is more than 10 items then it forwards the request to the next proxy in the sequence (i.e. the HighQuantityProxy).  The operation placeOrder can procced if the user is authenticated, so the LowQuantityProxy is this way is implementing the Proxy design pattern. 

HighQuantityProxy.java. This class implements the logic of the placeOrder operation by forwarding the request to the placeOrder operation of the Facade. The operation placeOrder can procced if the user is authenticated, so the HighQuantityProxy is this way is implementing the Proxy design pattern. There is no other proxy in the sequence after the HighQuantityProxy. You are free to change this order and implement your own proxies.